js/xpconnect/tests/chrome/test_doublewrappedcompartments.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=533596
-->
<window title="Mozilla Bug 533596"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">

  <iframe type="content"
          src="http://example.org/tests/js/xpconnect/tests/mochitest/file_doublewrappedcompartments.html"
          onload="go()"
          id="ifr">
  </iframe>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[
      const utils = SpecialPowers.DOMWindowUtils;

      function go() {
        var wrappedWin = $('ifr').contentWindow;
        is(typeof wrappedWin.expando, 'undefined', "Shouldn't be able to see the expando");

        var unwrapped = wrappedWin.wrappedJSObject;

        var expando = unwrapped.expando;
        isnot(expando, 'undefined', 'properly wrapped');
        is(typeof expando.querySelector, 'function', 'double wrapped');

        SimpleTest.finish();
      }

      SimpleTest.waitForExplicitFinish();
  ]]></script>
</window>
